# AtlasRoutes/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AtlasRoutes API",
        "description": "API for managing tourist itineraries, destinations, and user favorites.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/stats/users-by-month": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Get users registered per month",
                "operationId": "2090b2d2c103070e83e1459959fc41f8",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter by year (YYYY)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2026
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monthly user counts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StatsUsersByMonthRow"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StatsUsersByMonthRow": {
                "properties": {
                    "month": {
                        "type": "string",
                        "example": "2026-03"
                    },
                    "total": {
                        "type": "integer",
                        "example": 27
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}